Updated Blog service and controller
Updated BlogService to integrate with the new RecipeService, leveraging its updated methods for creating, updating, and deleting recipes.
Enhanced logic in CreateBlog and UpdateBlog methods to handle cases where a blog may or may not include a recipe.
Added authorization checks in BlogController for update-blog and delete-blog endpoints to ensure only blog owners or admins can modify or delete blogs.

+++ BackEngin/Controllers/BlogController.cs
@@ -1,25 +1,27 @@
-﻿using BackEngin.Services.Interfaces;
+﻿using Asp.Versioning;
+using BackEngin.Services.Interfaces;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Models.DTO;
 
 namespace BackEngin.Controllers
 {
+    [ApiVersion("1.0")]
+    [Route("api/v{version:apiVersion}/blog")]
     [ApiController]
-    [Route("api/[controller]")]
-    public class BlogController : ControllerBase
+    public class BlogController : ApiControllerBase
     {
         private readonly IBlogService _blogService;
 
-        public BlogController(IBlogService blogService)
+        public BlogController(IBlogService blogService) : base()
         {
             _blogService = blogService;
         }
 
         [HttpGet("blogs")]
-        public async Task<IActionResult> GetBlogs()
+        public async Task<IActionResult> GetBlogs([FromQuery] int pageNumber = 1, [FromQuery] int pageSize = 10)
         {
-            var blogs = await _blogService.GetBlogs();
+            var blogs = await _blogService.GetBlogs(pageNumber, pageSize);
             return Ok(blogs);
         }
 
@@ -29,6 +31,7 @@ namespace BackEngin.Controllers
         {
             if (createBlogDto == null) return BadRequest("Invalid blog data.");
 
+            createBlogDto.UserId = await GetActiveUserId();
             var createdBlog = await _blogService.CreateBlog(createBlogDto);
 
             return CreatedAtAction(nameof(GetBlogById), new { blogId = createdBlog.Id }, createdBlog);
@@ -46,19 +49,66 @@ namespace BackEngin.Controllers
         [Authorize]
         public async Task<IActionResult> UpdateBlog(int blogId, [FromBody] UpdateBlogDTO updateBlogDto)
         {
-            var updatedBlog = await _blogService.UpdateBlog(blogId, updateBlogDto);
-            if (updatedBlog == null) return NotFound();
-            return Ok(updatedBlog);
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            var blogOwner = await _blogService.GetOwner(blogId);
+            // Return NotFound if the blog doesn't exist
+            if (blogOwner == null)
+            {
+                return NotFound();
+            }
+
+            if (!await CanUserAccess(blogOwner))
+            {
+                return Unauthorized();
+            }
+
+            try
+            {
+                var updatedBlog = await _blogService.UpdateBlog(blogId, updateBlogDto);
+                if (updatedBlog == null) // Double-check if the update failed
+                {
+                    return NotFound();
+                }
+                return Ok(updatedBlog);
+            }
+            catch (ArgumentException ex)
+            {
+                return BadRequest(ex.Message);
+            }
+            catch (Exception ex)
+            {
+                // Log the exception details if necessary
+                return StatusCode(StatusCodes.Status500InternalServerError, "An unexpected error occurred.");
+            }
         }
+
 
         [HttpDelete("delete-blog/{blogId}")]
         [Authorize]
         public async Task<IActionResult> DeleteBlog(int blogId)
         {
+            var blogOwner = await _blogService.GetOwner(blogId);
+            if (blogOwner == null)
+            {
+                return NotFound();
+            }
+
+            if (!await CanUserAccess(blogOwner))
+            {
+                return Unauthorized();
+            }
+
             var result = await _blogService.DeleteBlog(blogId);
-            if (!result) return NotFound();
-            return NoContent();
+            if (!result)
+            {
+                return NotFound();
+            }
+
+            return Ok(new { message = "Blog deleted successfully!" });
         }
+
 
         [HttpGet("blog-recipe/{blogId}")]
         public async Task<IActionResult> GetRecipeOfBlog(int blogId)
